group and rename config related modules
