+++ flex-tasks/flex-tasks.cabal
@@ -12,8 +12,9 @@ extra-source-files:
 
 library
   exposed-modules:
+      FlexTask.Config.Default
+      FlexTask.Config.Types
       FlexTask.ConvertForm
-      FlexTask.DefaultConfig
       FlexTask.FormHelpers
       FlexTask.FormUtil
       FlexTask.Generic.Form
@@ -22,7 +23,6 @@ library
       FlexTask.Interpreter.EvaluationHelper
       FlexTask.Interpreter.GenerationHelper
       FlexTask.Interpreter
-      FlexTask.Types
       FlexTask.YesodConfig
   other-modules:
       FlexTask.FormTypes
@@ -131,12 +131,12 @@ test-suite flex-tasks-test
   type: exitcode-stdio-1.0
   main-is: Spec.hs
   other-modules:
-      FlexTask.DefaultConfigSpec
+      FlexTask.Config.DefaultSpec
+      FlexTask.Config.TypesSpec
       FlexTask.FormUtilSpec
       FlexTask.Generic.FormSpec
       FlexTask.Generic.ParseSpec
       FlexTask.TestUtil
-      FlexTask.TypesSpec
       Paths_flex_tasks
   hs-source-dirs:
       test
