+++ flex-tasks/src/FlexTask/Interpreter.hs
@@ -54,7 +54,7 @@ import System.Directory (
 import System.Environment          (getEnv)
 import System.FilePath             ((</>), (<.>))
 
-import FlexTask.Types (
+import FlexTask.Config.Types (
   CommonModules(..),
   FlexConf(..),
   FlexInst(..),
